Remove obsolete EQs attached to some of the patternized terms
This commit intends to remove some old equivalence axioms attached to some term definitions that use the `AbnormalMorphologyOfPartOfAnatomicalEntity` pattern template.
These terms used to use be defined by the `abnormalMorphologyOfAnatomicalEntity` template, but they are now switched to using `AbnormalMorphologyOfPartOfAnatomicalEntity`.
This commit intends to remove the obsolete axioms related to the `abnormalMorphologyOfAnatomicalEntity` template.